+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/vite-plugin
 
+## 0.44.0-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.44.0-pre.0
+
 ## 0.43.1
 
 ### Patch Changes